# doc/api.yaml
openapi: 3.0.0
info:
  title: Fantastic Coffee (decaffeinated)
  version: 1.0.0
servers:
  - url: 'https://localhost:3000'
...
components:
  schemas:
    UserID:
      description: The unique id of a user, automatically generated when the account is created.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
    UserName:
      description: The username; it can be modified, but no 2 people can have the same one at the same time.
      type: string
      pattern: 'ˆ.*?$'
      minLength: 3
      maxLength: 16
    Image:
      description: The image object, made of its name and the actual image base64-encoded.
      type: string
      format: byte
      pattern: 'ˆ.*?$'
      minLength: 1
      maxLength: 10000000
    MessageID:
      description: The unique identifier of a message, automatically generated when the message is first created and sent.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
    Text:
      description: The content of a simple/plain text messsage.
      type: string
      pattern: 'ˆ.*?$'
      minLength: 1
      maxLength: 10000
    ConversationID:
      description: The unique identifier of a chat, automatically generated when the conversation is started.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
    ConversationName:
      description: The name of the conversation, which can be either the name of a group or of a person.
      type: string
      pattern: 'ˆ.*?$'
      minLength: 1
      maxLength: 20
    TimeStamp:
      description: The time at which the last message was sent.
      type: string
      format: date-time
      minLength: 20
      maxLength: 20
      readOnly: true
      example: "2017-07-21T17:32:28Z"
...
    Emoji:
      description: an emoji
      type: string
      pattern: "[\u203C-\u3299\u1F004\u1F0CF\u1F9FF]"
      minLength: 0
      maxLength: 1
      default: ""
    Reaction: 
      description: The reaction/emoji to a messsage; each person can put at most one emoji.
      type: object
      properties:
        sender:
          $ref: "#/components/schemas/UserID"
        emoji:
          $ref: "#/components/schemas/Emoji"
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UserID"
        name:
          $ref: "#/components/schemas/UserName"
        image:
          $ref: "#/components/schemas/Image"
    MessageContent:
      type: object
      properties:
        text:
          $ref: "#/components/schemas/Text"
        image:
          $ref: "#/components/schemas/Image"
    Message:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/MessageID"
        conversation:
          $ref: "#/components/schemas/ConversationID"
        sender:
          $ref: "#/components/schemas/UserID"
        timestamp:
          $ref: "#/components/schemas/TimeStamp"
        checkmark:
          description: Tells you whether the message was received, read, or none of the two.
          type: string
          format: byte
          enum:
            - "sent"
            - "received"
            -  "read"
        text:
          $ref: "#/components/schemas/Text"
        image:
          $ref: "#/components/schemas/Image"
        responseTo:
          $ref: "#/components/schemas/MessageID"
        reactions:
          description: The list of all emojis (reactions) to this message.
          type: array
          minItems: 0
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Reaction"
    Conversation:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ConversationID"
        name:
          $ref: "#/components/schemas/ConversationName"
        image:
          $ref: "#/components/schemas/Image"
        group:
          description: flag indicating whether the conversation is a group (true) or (false)
          type: boolean
          default: true
          readOnly: true
        lastmessage:
          $ref: "#/components/schemas/MessageID"
        participants:
          description: List of all members of a group chat. 
          type: array
          minItems: 3
          maxItems: 50
          items:
            $ref: "#/components/schemas/User"
          nullable: true
        messages:
          description: The list of all messages sent of the chat.
          type: array
          minItems: 0
          maxItems: 100000
          items:
            $ref: "#/components/schemas/Message"
  parameters:
    SearchName:
      description: The name string used to search for users or conversations
      name: srcName
      in: query
      required: false
      schema:
        type: string
        format: byte
        pattern: 'ˆ.*?$'
        minLength: 0
        maxLength: 20
        default: ""
    UserId:
      description: The user identifier, which must be passed for any possible operation.
      name: usrId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserID"
    ConversationId:
      description: The identifer of a conversation, used for any kind of acess to that conversation.
      name: convId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ConversationID"
...
  securitySchemes:

    AuthenticationId:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid.
    ServerError: 
      description: Server error occurred. Please, try again in a few minutes.
    NotFoundError:
      description: The given id(s) doesn't correspond witha any user/session/message. Please try again.
    BadRequestError:
      description: You either aren't an admin to do this operation or the given input(s) are invalid.
    ItemDeleted:
      description: Item(s) succesfully deleted.
...
security:
  - AuthenticationId: []
paths:
...
  /users/{usrId}/others:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [user]
      summary: let you search all users
      description: |-
        allows you to search other users via the username. 
        if more than one match is found, only the first 50 are shown.
      operationId: getUsers
      parameters:
        - $ref: "#/components/parameters/SearchName"
      responses:
        "200": 
          description: The following users have been found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    minItems: 1
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{usrId}/name:
    parameters:
      - $ref: "#/components/parameters/UserId"
    put:
      tags: [user]
      summary: modifies the username, if possible
      description: |-
        if the chosen username isn't taken, modifies it, with the one specified.
        Then, it returns the updated profile of the user.
      operationId: setMyUserName
      requestBody:
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/UserName"
      responses:
        "200":
          description: User name modified successfully
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/UserName"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{usrId}/conversations/{convId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [conversation]
      summary: opens the chosen conversation
      description: |-
        based on the given id, taken from the conversation's preview,
        it returns the corresponding conversation
      operationId: getConversation
      responses:
        "200":
          description: conversation successfully loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
    post: 
      tags: [group]
      summary: add new members to group
      description: |-
        adds the chosen participant to the group chat, then returns the conversation,
        with the newly updated list of participants.
      operationId: addToGroup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                participants:
                  type: array
                  minItems: 3
                  maxItems: 1000
                  items:
                    $ref: "#/components/schemas/UserName"
      responses:
        "201":
          description: New members succesfully added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  participants:
                    type: array
                    minItems: 3
                    maxItems: 1000
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [group]
      summary: leaves the group
      description: |-
        makes you leave a conversation, completely removing you from the group participants' list.
        the group will also be removed from your list of conversations and its messages deleted.
      operationId: leaveGroup
      responses:
        "204":
          $ref: "#/components/responses/ItemDeleted"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{usrId}/conversations/{convId}/name:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
    put:
      tags: [group]
      summary: changes the group's name
      description: |-
        if the conversation is a group chat, modifies its name with the one given in the request body,
        then returns the modified status of the conversation;
        otherwise, the name remains unchanged
      operationId: setGroupName
      requestBody:
        content:
          text/plain:
            schema: 
              $ref: "#/components/schemas/ConversationName"
      responses:
        "200":
          description: Group name modified successfully
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ConversationName"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{usrId}/conversations/{convId}/image:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
    put:
      tags: [group]
      summary: changes the group's image
      description: |-
        modifies the chat's name with the one given in the request body,
        then returns the modified status of the conversation;
        otherwise, the picture remains unchanged
      operationId: setGroupPhoto
      requestBody:
        content:
          application/json:
              schema:
                type: object
                properties:
                  image:
                    $ref: "#/components/schemas/Image"
      responses:
        "200":
          description: User image modified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  image:
                    $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{usrId}/conversations/{convId}/messages:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
    post: 
      tags: [message]
      summary: sends new message
      description: |-
        creates a new message and sends it to the conversation, adding it to the conversation's log upon success,
        if the new message has the "responseTo" property != "", it means
        this new message is an answer to the message specified in the "response" property
      operationId: sendMessage
      parameters:
      - name: responseTo
        in: query
        required: false
        schema:
          $ref: "#/components/schemas/MessageID"
      requestBody:
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/MessageContent"
      responses:
        "201":
          description: New message sent succesfully.
          content:
            application/json:
              schema: 
                  $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
...